utils/jwt_utils.py
import jwt
from core.config import config
from passlib.context import CryptContext
from datetime import datetime, timedelta, timezone
from utils.handle_error import ErrorHandler
from fastapi.security import OAuth2PasswordBearer
from jwt.exceptions import InvalidTokenError

# 密码哈希配置
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")

# 工具函数
def get_password_hash(password: str):
    return pwd_context.hash(password)

def verify_password(plain_password: str, hashed_password: str):
    return pwd_context.verify(plain_password, hashed_password)

def create_access_token(data: dict, expires_delta: timedelta | None= None):
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.now(timezone.utc) + expires_delta
    else:
        expire = datetime.now(timezone.utc) + timedelta(minutes=15)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, config.SECRET_KEY, algorithm=config.ALGORITHM)
    return encoded_jwt

def verify_access_token(token: str):
    try:
        payload = jwt.decode(token, config.SECRET_KEY, algorithms=config.ALGORITHM)
        exp = payload.get("exp")
        if exp and datetime.now(timezone.utc) > datetime.fromtimestamp(exp, timezone.utc):
            ErrorHandler.handle_token_timeout()
        return payload

    except InvalidTokenError :
        ErrorHandler.handle_token_invaliable()