requirements.txt
uvicorn
python-dotenv
dnspython
fastapi
starlette
motor
pydantic
pydantic-settings
passlib[bcrypt]
pyjwt
bcrypt==4.0.1